Split compression related tests into a test group #3689
