+++ pom.xml
@@ -841,6 +841,7 @@
                                         <sortModulesPath>integration-test-groups/aws2/pom.xml</sortModulesPath>
                                         <sortModulesPath>integration-test-groups/aws2-quarkus-client/pom.xml</sortModulesPath>
                                         <sortModulesPath>integration-test-groups/azure/pom.xml</sortModulesPath>
+                                        <sortModulesPath>integration-test-groups/compression/pom.xml</sortModulesPath>
                                         <sortModulesPath>integration-test-groups/foundation/pom.xml</sortModulesPath>
                                         <sortModulesPath>integration-test-groups/mongodb/pom.xml</sortModulesPath>
                                     </sortModulesPaths>
@@ -899,6 +900,15 @@
                                                 </includes>
                                             </sourcePoms>
                                             <destinationPom>${basedir}/integration-tests/azure-grouped/pom.xml</destinationPom>
+                                        </mergePom>
+                                        <mergePom>
+                                            <sourcePoms>
+                                                <directory>${basedir}/integration-test-groups/compression</directory>
+                                                <includes>
+                                                    <include>*/pom.xml</include>
+                                                </includes>
+                                            </sourcePoms>
+                                            <destinationPom>${basedir}/integration-tests/compression-grouped/pom.xml</destinationPom>
                                         </mergePom>
                                         <mergePom>
                                             <sourcePoms>
